Made menu fully functional including finishing a meal

+++ app/src/main/java/com/instamealmobile/data/Recipe.kt
@@ -20,5 +20,5 @@ data class Recipe(
 data class Record(
     val household_id: String,
     val timestamp: String,
-    val rating: Int?
+    val rating: Float?
 )
